Fix state bug and `on()` default
I always forget - you can't send a deep object to hue. It complains that
`state` isn't a valid value. The reason is that you do a `put` to
`${light}/state`, not the light itself. Pretty depressing, cuz otherwise
you could manage an entire light through a single object. Oh well...
So this commit removes the `state` object, instead treating it as a
single-level object. It's been confirmed to work with the lights.
Also, `.on()` without parameters would turn the light off. That was
because there was no default, so off (being falsy) ruled. That's been
fixed and tested for now.