Ignore troublesome state keys
It's bad if you send a state to the hue bridge specifying three
different color updates. However, when you use `State.from` to import a
hue state object, all those properties will be sent to the bridge, even
though you might've only changed hsl. This confuses the bridge and
usually ends badly. Since a state object is only meant to be written to,
not read, it should be safe to filter out some of the unwanted input
(like `reachable`, `colormode`, `xy`, etc.).